[refactor] created features/auth and  moved auth files in it.
