+++ src/app/features/index.ts
@@ -1,6 +1,6 @@
-export * from './authSlice';
-export * from './userSlice';
-export * from './tweetSlice';
-export * from './bookmarkSlice';
-export * from './thunkApiCalls/tweetThunk';
-export * from './thunkApiCalls/authThunk';
+export * from "./Auth/authSlice";
+export * from "./Auth/authThunk";
+export * from "./userSlice";
+export * from "./tweetSlice";
+export * from "./bookmarkSlice";
+export * from "./thunkApiCalls/tweetThunk";
