+++ src/app/store.ts
@@ -1,17 +1,17 @@
 import { TypedUseSelectorHook, useDispatch, useSelector } from "react-redux";
 import { configureStore } from "@reduxjs/toolkit";
-import auth from './features/authSlice';
-import users from './features/userSlice';
-import tweets from './features/tweetSlice';
-import bookmarks from './features/bookmarkSlice';
+import auth from "./features/Auth/authSlice";
+import users from "./features/userSlice";
+import tweets from "./features/tweetSlice";
+import bookmarks from "./features/bookmarkSlice";
 
 export const store = configureStore({
-	reducer: {
-		auth,
-		users,
-		tweets,
-		bookmarks
-	}
+  reducer: {
+    auth,
+    users,
+    tweets,
+    bookmarks,
+  },
 });
 
 // Infer the `RootState` and `AppDispatch` types from the store itself
